ForgetNotBackend/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-cors-headers==3.14.0
django-redis==5.2.0
djangorestframework==3.14.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
PyMySQL==1.0.2
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
